sqlite_to_postgres/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-split-settings==1.2.0
flake8==6.1.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
